// app/src/app/shared/models/data-service-gtp.ts
export interface DataServiceGTP {
  id: number;
  name: string;
  debtorCode: string;
  dataType: string;
  paymentType: string;
  idAccount: string;
  accountNumber: string;
  currency: string;
  useAppWeb: boolean;
  useAgent: boolean;
  useStore: boolean;
  partialPayment: string;
  chargeInterest: string;
  chargeType: number | string;
  interestType: string;
  amount: number;
  porcentage: number;
  currencySymbol: string;
  inReview: boolean;
  newNameCode: string;
  newName: string;
  status: string;
  // newNameGtpStatus?: number;
  // newNameCodeGtpStatus?: number;
  newNameGTPStatus: number;
  newNameCodeGTPStatus: number;
  acceptednewNameCode?: boolean;
  acceptednewName?: boolean;
  nombreHabilitado?: boolean;
  nombreCodHabilitado?: boolean;

  res?: string;
  nombre?: string;
  tipoDato?: string;
  cobraMora?: string;
  tipoMora?: string;
  simboloMoneda?: string;
  porcentaje?: number;
  monto?: number;
  nroCuenta?: string;
  tipoPago?: string;
  rubro?: string;
  codDeudor?: string;
  nameCod?: string;
  moneda?: string;
  usaWebApp?: boolean;
  usaAgente?: boolean;
  usaTienda?: boolean;
  periodoMora?: string;
  pagoPartes?: string;
  useAgencyChannel?: boolean;
}